+++ src/safety_zone/center_control.cpp
@@ -19,33 +19,50 @@ CenterControl::CenterControl(Prism& prism, std::string frame_id, ros::NodeHandle
   init();
 }
 
-CenterControl::CenterControl(SafetyZone& safety_zone, int obstacle_id, std::string frame_id, ros::NodeHandle nh) : id(id_generator), prism_(safety_zone.getObstacle(obstacle_id)) {
+CenterControl::CenterControl(SafetyZone* safety_zone, int obstacle_id, std::string frame_id, ros::NodeHandle nh) : id(id_generator), obstacle_id_(obstacle_id), prism_(safety_zone->getObstacle(obstacle_id)) {
+  frame_id_ = frame_id;
+  nh_ = nh;
+  id_generator++;
+  safety_zone_ = safety_zone;
+  init();
+}
+
+CenterControl::CenterControl(SafetyZone* safety_zone, std::string frame_id, ros::NodeHandle nh) : id(id_generator), prism_(safety_zone->getBorder()){
   frame_id_ = frame_id;
   nh_ = nh;
   id_generator++;
   init();
 }
 
-CenterControl::CenterControl(SafetyZone& safety_zone, std::string frame_id, ros::NodeHandle nh) : id(id_generator), prism_(safety_zone.getBorder()) {
-  frame_id_ = frame_id;
-  nh_ = nh;
-  id_generator++;
-  init();
+CenterControl::~CenterControl(){
+  if(server_){
+    delete server_;
+  }
+  if(menu_handler_){
+    delete menu_handler_;
+  }
+  if(is_active_){
+    prism_.unsubscribe(this);
+  }
+  cleanup();
 }
+
 
 void CenterControl::init(){
   server_ = new interactive_markers::InteractiveMarkerServer(nh_.getNamespace() + "safety_area_center_out", std::to_string(id), false);
   menu_handler_ = new interactive_markers::MenuHandler();
-  menu_handler_->insert("Delete the prism", [this](const vm::InteractiveMarkerFeedbackConstPtr &feedback){this->deleteCallback(feedback);});
+  if(safety_zone_){
+    menu_handler_->insert("Delete the prism", [this](const vm::InteractiveMarkerFeedbackConstPtr &feedback){this->deleteCallback(feedback);});
+  }else{
+    menu_handler_->insert("Deleting is disabled for this prism");
+  }
 
   prism_.subscribe(this);
   addIntMarker();
 }
 
 void CenterControl::update(){
-  if(!prism_.isActive()){
-    server_->clear();
-    server_->applyChanges();
+  if(!is_active_){
     return;
   }
 
@@ -58,6 +75,11 @@ void CenterControl::update(){
   server_->applyChanges();
 }
 
+void CenterControl::cleanup(){
+  server_->clear();
+  server_->applyChanges();
+  is_active_ = false;
+}
 
 vm::Marker CenterControl::makeBox(vm::InteractiveMarker &msg){
   vm::Marker marker;
@@ -164,7 +186,7 @@ void CenterControl::mouseUpCallback(const visualization_msgs::InteractiveMarkerF
 }
 
 void CenterControl::deleteCallback(const visualization_msgs::InteractiveMarkerFeedbackConstPtr &feedback){
-  prism_.deactivate();
+  safety_zone_->deleteObstacle(obstacle_id_);
 }
